refactor: Use try-with-resources in FileUtil.copyFile
This change refactors the `copyFile` method in `FileUtil.java` to use a try-with-resources statement for managing `FileInputStream` and `FileOutputStream`. This ensures that the streams are properly closed, even if an exception occurs during the copy operation. Additionally, error logging has been added to provide more information in case of an `IOException`.